refactor(centrals): share HS256 verify primitive; drop presigned URL from startRestore
Follow-up on two deferred review items:
- HS256 dedup: extract the shared verifyHs256Signature primitive (constant-time HMAC + payload decode) into shared/utils/jwtHs256; AuthService.verifyJWT and centralAuth both use it, keeping their own (different) claim validation. The two verifiers can no longer silently diverge.
- startRestore no longer mints/returns a presigned download URL to the browser (F-B3): the central fetches the dump via its own agent poll (nextJob). Removes the now-unused storage dependency from CentralRestoreService.
